feat: improve exec helpers
allow to get container and outputs
